Allow empty invalid Syl instance
Syl instance constructed with default c-tor should be allowed,
so incremental construction during parsing is possible.
For this, the mandatory parts of the syllable must be INVALID,
namely the initial consonant (and its class), the vowel, and the tone.
Other optional components (i.e. second consonant, ending consonant)
can be initialized as NONE.

* sylstring/sylcomp.h:
  - Add INVALID value to EInitCons, EVowel, ETone
* thortho/thcons.h:
  - Add INVALID value to EInitConsClass
* sylstring/sylstring.h, sylstring/sylstring.inl:
  - Add Syl default c-tor